+++ libpartitioner/include/storage/DataReader.h
@@ -27,8 +27,8 @@ class DataReader {
         arrow::Result<std::shared_ptr<arrow::ChunkedArray>> getColumn(const std::string &columnName);
         void displayFileProperties();
         arrow::Result<std::pair<double_t, double_t>> getColumnStats(const std::string &columnName);
-        uint32_t getNumRows();
-        uint32_t getExpectedNumBatches();
+        int64_t getNumRows();
+        int64_t getExpectedNumBatches();
         static arrow::Result<std::shared_ptr<arrow::Table>> getTable(std::filesystem::path &inputFile);
         static std::filesystem::path getDatasetPath(const std::filesystem::path &folder, const std::string &datasetName,
                                                     const std::string &partitioningScheme);
